+++ app/Models/User.php
@@ -26,6 +26,11 @@ class User extends Authenticatable
         'password',
         'theme_slug',
         'timezone',
+        'ui_font',
+        'manuscript_font',
+        'ui_scale',
+        'manuscript_scale',
+        'manuscript_leading',
     ];
 
     /**